新增分布式锁

+++ src/main/java/com/jaffer/btrip/service/impl/CorpServiceImpl.java
@@ -5,6 +5,7 @@ import com.jaffer.btrip.manager.CorpManager;
 import com.jaffer.btrip.service.CorpService;
 import com.jaffer.btrip.util.BtripResult;
 import com.jaffer.btrip.util.BtripResultUtils;
+import com.jaffer.btrip.util.RedisLockUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.BooleanUtils;
 import org.springframework.stereotype.Service;
@@ -15,12 +16,25 @@ import java.util.Objects;
 @Service
 @Slf4j
 public class CorpServiceImpl implements CorpService {
+
+
+    private static final String LOCK_KEY_REGISTER_CORP = "LOCK_KEY_REGISTER_CORP_%s";
+
+    private static final String LOCK_KEY_MAINTAIN_CORP = "LOCK_KEY_MAINTAIN_CORP_%s";
+
+
     @Resource
     private CorpManager corpManager;
 
     @Override
     public BtripResult<Boolean> registerCorp(String corpName, String phoneNumber, String userName) {
+        String lockKey = String.format(LOCK_KEY_REGISTER_CORP, corpName);
         try {
+            boolean lock = RedisLockUtils.tryLock(lockKey);
+            if (BooleanUtils.isFalse(lock)) {
+                return BtripResultUtils.returnFailMsg("重复注册，请稍后重试");
+            }
+
             Boolean registerCorp = corpManager.registerCorp(corpName, phoneNumber, userName);
             if (BooleanUtils.isFalse(registerCorp)) {
                 return BtripResultUtils.returnFailMsg("创建企业失败");
@@ -29,6 +43,8 @@ public class CorpServiceImpl implements CorpService {
         } catch (Exception e) {
             log.error("registerCorp error, corpName:{}, phoneNumber:{}, userName:{}", corpName, phoneNumber, userName,e);
             return BtripResultUtils.returnFailMsg("创建企业失败,失败原因 :" + e.getMessage());
+        } finally {
+            RedisLockUtils.unlock(lockKey);
         }
     }
 
@@ -49,7 +65,12 @@ public class CorpServiceImpl implements CorpService {
 
     @Override
     public BtripResult<Boolean> deleteCorpByCorpId(String corpId) {
+        String lockKey = String.format(LOCK_KEY_MAINTAIN_CORP, corpId);
         try {
+            boolean lock = RedisLockUtils.tryLock(lockKey);
+            if (BooleanUtils.isFalse(lock)) {
+                return BtripResultUtils.returnFailMsg("多个用户正在维护企业信息，请稍后重试");
+            }
 
             Boolean deleteRes = corpManager.deleteCorpByCorpId(corpId);
 
@@ -61,6 +82,8 @@ public class CorpServiceImpl implements CorpService {
         } catch (Exception e) {
             log.error("deleteCorpByCorpId error, corpId:{}", corpId,e);
             return BtripResultUtils.returnFailMsg("删除企业失败,失败原因 :" + e.getMessage());
+        } finally {
+            RedisLockUtils.unlock(lockKey);
         }
     }
 }
